Complete integration of rust logger with OSL
