+++ Cargo.toml
@@ -11,7 +11,20 @@ name = "opensourceleg_rs"
 crate-type = ["cdylib"]
 
 [dependencies]
-pyo3 = { version = "0.25", features = ["extension-module"] }
 numpy = "0.25"
 nalgebra = "0.33"
-serde = { version = "1.0", features = ["derive"] }
+chrono = {version = "0.4.41", features = ["serde"]}
+dashmap = { version = "6.1.0", features = ["serde"] }
+once_cell = "1.21.3"
+serde = {version="1.0.219", features=["derive"]}
+serde_json = "1.0.140"
+pyo3 = { version = "0.25.1", features = ["extension-module"]}
+tracing = "0.1.41"
+tracing-appender = "0.2.3"
+tracing-subscriber = { version = "0.3.19", features = ["chrono", "env-filter"] }
+valuable = "0.1.1"
+pythonize = "0.25.0"
+zenoh = "1.4.0"
+tokio = { version = "1.46.1", features = ["full"] }
+fxhash = "0.2.1"
+cdr = "0.2.4"
