--- opensourceleg/logging/observe/.gitignore
@@ -1,3 +1,0 @@
-/target
-/logs
-/benchmark/__pycache__
