--- opensourceleg/logging/observe/config.alloy
@@ -1,21 +1,0 @@
-loki.write "local" {
-  endpoint {
-    url = "http://loki:3100/loki/api/v1/push"
-  }
-}
-
-loki.source.file "json_logs" {
-  targets = [{
-    __path__ = "/logs/variables.log",
-  }]
-  forward_to = [loki.write.local.receiver]
-}
-
-local.file_match "plain_logs_match" {
-  path_targets = [{"__path__" = "/logs/**/logfile*.log"}]
-}
-
-loki.source.file "plain_logs" {
-  targets = local.file_match.plain_logs_match.targets
-  forward_to = [loki.write.local.receiver]
-}
