--- opensourceleg/logging/observe/docker-compose.yml
@@ -1,32 +1,0 @@
-#docker run ^  -v "%cd%\config.alloy:/etc/alloy/config.alloy" ^ -v "%cd%\logs:/logs"^  -p 12345:12345 ^  grafana/alloy:latest ^    run --server.http.listen-addr=0.0.0.0:12345 --storage.path=/var/lib/alloy/data ^    /etc/alloy/config.alloy
-
-services:
-  grafana:
-    image: grafana/grafana-oss
-    ports:
-      - "3000:3000"
-    volumes:
-      - grafana-storage:/var/lib/grafana
-    depends_on:
-      - loki
-
-  loki:
-    image: grafana/loki:latest
-    ports:
-      - "3100:3100"
-    command: -config.file=/etc/loki/local-config.yaml
-
-  alloy:
-    image: grafana/alloy:latest
-    depends_on:
-      - loki
-    ports:
-      - "12345:12345"
-    volumes:
-      - ./config.alloy:/etc/alloy/config.alloy
-      - ./logs:/logs:ro
-    command:
-      run --server.http.listen-addr=0.0.0.0:12345 --storage.path=/var/lib/alloy/data /etc/alloy/config.alloy
-
-volumes:
-  grafana-storage:
