--- opensourceleg/logging/observe/zenoh.json5
@@ -1,5 +1,0 @@
-{
-  connect: {
-    endpoints: ["tcp/192.168.68.125:7447"]
-  }
-}
