+++ opensourceleg_rs.pyi
@@ -120,13 +120,6 @@ class Logger:
         """
 
     @staticmethod
-    def start_ros_subscriber(config_file: str, key_expr: str) -> None:
-        """
-        Starts a background thread which subscribes to the channel
-        defined by key_expr parameter.
-        """
-
-    @staticmethod
     def set_console_level(log_level: LogLevel) -> None:
         """
         Changes the level of log messages output to console at runtime.
