+++ src/lib.rs
@@ -1,10 +1,15 @@
 use pyo3::prelude::*;
-
+use crate::logger::{Logger, PyLogLevel};
+pub mod logger;
+pub mod record;
+pub mod rotator;
+pub mod profiler;
+pub mod core;
 mod math;
 
 /// High-performance math operations for robotics
 #[pymodule]
-fn rust(m: &Bound<'_, PyModule>) -> PyResult<()> {
+fn opensourceleg_rs(m: &Bound<'_, PyModule>) -> PyResult<()> {
     m.add("__version__", "3.1.0")?;
 
     // Add math submodule
@@ -12,5 +17,27 @@ fn rust(m: &Bound<'_, PyModule>) -> PyResult<()> {
     math::register_module(&math_module)?;
     m.add_submodule(&math_module)?;
 
+    m.add_class::<Logger>()?;
+    m.add_class::<PyLogLevel>()?;
+    m.add_class::<profiler::PyProfiler>()?;
+
+    let atexit = m.py().import("atexit")?;
+    let logger_class = m.getattr("Logger")?;
+    let flush_all_method = logger_class.getattr("flush_all")?;
+    atexit.call_method1("register", (flush_all_method,))?;
+
     Ok(())
 }
+
+// #[pymodule]
+// fn observable(py: Python, m: &Bound<'_, PyModule>) -> PyResult<()>{
+//     m.add_class::<Logger>()?;
+//     m.add_class::<PyLogLevel>()?;
+//     m.add_class::<profiler::PyProfiler>()?;
+
+//     let atexit = py.import("atexit")?;
+//     let logger_class = m.getattr("Logger")?;
+//     let flush_all_method = logger_class.getattr("flush_all")?;
+//     atexit.call_method1("register", (flush_all_method,))?;
+//     Ok(())
+// }
